requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
asyncpg>=0.30.0

# Web scraping
crawl4ai>=0.4.0
beautifulsoup4>=4.12.0
httpx>=0.26.0
lxml>=5.1.0

# AI/ML
anthropic>=0.39.0

# Email
mailchimp-marketing>=3.0.80
jinja2>=3.1.3

# Scheduler
apscheduler>=3.10.4

# Configuration
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1

# Development/Testing (optional)
pytest>=7.4.4
pytest-asyncio>=0.23.3
